requirements.txt
aiohttp
appdirs
asgiref
async-timeout
attrs
black
cachetools
certifi
cffi
chardet
charset-normalizer
click
coreapi
coreschema
Django
django-debug-toolbar
django-ninja
django-ratelimit
django-seed
django-storages
django-user-agents
djangorestframework
drf-yasg
Faker
geoip2
google-api-core
google-auth
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
gunicorn
idna
inflection
itypes
Jinja2
MarkupSafe
maxminddb
multidict
mypy-extensions
mysqlclient
packaging
pathspec
protobuf
pyasn1
pyasn1-modules
pycparser
pydantic
pyparsing
python-dateutil
pytz
regex
requests
rsa
ruamel.yaml
ruamel.yaml.clib
six
sqlparse
text-unidecode
toml
typing-extensions
ua-parser
uritemplate
urllib3
user-agents
yarl